requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.2
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==13.4.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.1
pytest-cov==3.0.0
pytest-deadfixtures==2.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
